moved stratified clustering out of main msm_we.msm_we into module
